requirements.txt
azure-identity
requests
click
azure-storage-blob